chore(TimePatternConverter) fixes package for TimePatternConverter. Converter has been moved from org.apache.camel.converter to org.apache.camel.catalog.impl
